requirements.txt
matplotlib==3.9.2
numpy==1.26.4
onnx==1.16.1
onnxruntime-gpu==1.19.2
onnxslim>=0.1.31
opencv-python==4.10.0.84
pandas==2.2.2
pillow==10.4.0
tensorboard==2.17.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
ultralytics==8.4.33
